requirements.txt
dnspython==1.15.0
pyaml==16.12.2
